% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{closeProject}
\alias{closeProject}
\title{Close Project}
\usage{
closeProject(simplace)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}
}
\description{
Call to the shutDown method of the simulation.
}
\seealso{
\code{\link{openProject}}
}
