% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{createSimulation}
\alias{createSimulation}
\title{Creates a simulation and substitute parameters}
\usage{
createSimulation(simplace, parameterList = NULL, queue = FALSE)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}

\item{parameterList}{a list with the parameter name as key and parametervalue as value}

\item{queue}{boolean - simulation is added to queue if true, else start new queue}
}
\value{
id of the created simulation
}
\description{
Creates a simulation from the opened project and substitutes
the values of the parameters given in the parameter list. 
Simulation will be put into the queue by default.
}
\seealso{
\code{\link{runSimulations}}, \code{\link{resetSimulationQueue}}
}
