% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{getDatatypesOfResult}
\alias{getDatatypesOfResult}
\title{Get the datatypes of the result variables}
\usage{
getDatatypesOfResult(result)
}
\arguments{
\item{result}{handle to the data container returned by \code{\link{getResult}}}
}
\value{
named character vector with the datatypes
}
\description{
Get the datatypes of each variable (i.e. data column).
The output is a named character vector, where each element is named
by the variables name.
}
