% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{getResult}
\alias{getResult}
\title{Fetch output from a simulation}
\usage{
getResult(simplace, outputId, simulationId = nullString)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}

\item{outputId}{id of the output. Only MEMORY outputs are accessible.}

\item{simulationId}{id of the simulation}
}
\value{
handle to the data container which has to be processed afterwards
}
\description{
The output is a JavaObject containing the variable names, 
data types, units and the values. Output can be converted
with \code{\link{resultToList}} or \code{\link{resultToDataframe}}
to R objects. Only MEMORY outputs are accessible. For CSV or 
database outputs you have to read the data by generic methods.
}
