% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{getSimplaceDirectories}
\alias{getSimplaceDirectories}
\title{Get the directories (work-, output-, projects- and data-dir)}
\usage{
getSimplaceDirectories(simplace)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}
}
\value{
character vector with the directories
}
\description{
Get the directories (work-, output-, projects- and data-dir)
}
\seealso{
\code{\link{setSimplaceDirectories}}
}
