% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{getSimulationIDs}
\alias{getSimulationIDs}
\title{Lists IDs of the performed simulations}
\usage{
getSimulationIDs(simplace)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object}
}
\value{
character vector with the IDs
}
\description{
Returns a vector with the IDs of the simulations. IDs are required to
get the output of the simulations.
}
