% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{getUnitsOfResult}
\alias{getUnitsOfResult}
\title{Get the units of the result variables}
\usage{
getUnitsOfResult(result)
}
\arguments{
\item{result}{handle to the data container returned by \code{\link{getResult}}}
}
\value{
character vector with the units
}
\description{
Get the units of each variable (i.e. data column) in a 
human readable format.
The output is a named character vector, where each element is named
by the variables name.
}
