% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapperHelper.R
\name{initSimplaceDefault}
\alias{initSimplaceDefault}
\title{Initialises Simplace with work- and outputdir for different settings}
\usage{
initSimplaceDefault(setting = "run")
}
\arguments{
\item{setting}{one of "run", "modules"," lapclient" or "wininstall"}
}
\value{
handle to the SimplaceWrapper object
}
\description{
Initialises Simplace with work- and outputdir for different settings
}
\seealso{
\code{\link{initSimplace}}
}
