% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{parameterListsToStringArray}
\alias{parameterListsToStringArray}
\title{Converts a list of named parameters to a java string array}
\usage{
parameterListsToStringArray(parameterLists)
}
\arguments{
\item{parameterLists}{list of parameter list (values indexed by parameter name)}
}
\value{
a java object of type Object[k][n][2]
}
\description{
Converts a list of named parameters to a java string array
}
\keyword{internal}
