% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{parameterListToStringArray}
\alias{parameterListToStringArray}
\title{Converts a list of named parameters to a java string array}
\usage{
parameterListToStringArray(parameterList)
}
\arguments{
\item{parameterList}{list of parameter values indexed by parameter name}
}
\value{
a java object of type Object[n][2]
}
\description{
Converts a list of named parameters to a java string array
}
\keyword{internal}
