% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{resetSimulationQueue}
\alias{resetSimulationQueue}
\title{Clears the list of simulations}
\usage{
resetSimulationQueue(simplace)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}
}
\value{
No return value, called for the side effect of clearing the simulation list
}
\description{
Simulation list is cleared
}
\seealso{
\code{\link{createSimulation}}, \code{\link{runSimulations}}
}
