% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{runProject}
\alias{runProject}
\title{Runs the opened project}
\usage{
runProject(simplace)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}
}
\value{
No return value, called for the side effect of running opened project
}
\description{
Runs the simulation(s) as defined in the solution and project files. 
There is no accessible MEMORY output, but one can load the CSV or 
database output.
}
\examples{
\dontrun{
simplace <- initSimplace(SimplaceInstallationDir,SimplaceWorkDir,SimplaceOutputDir)
openProject(simplace, Solution, Project)
runProject(simplace)
...
closeProject(simplace)   }
}
