% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{setProjectLines}
\alias{setProjectLines}
\title{Sets the lines of the project data files that should be used
when running a project.}
\usage{
setProjectLines(simplace, lines)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}

\item{lines}{either a vector of integers or a string of numbers separated by commas}
}
\value{
No return value, called for the side effect of selecting project to be run
}
\description{
You have to call the function after \code{\link{initSimplace}}
but before \code{\link{openProject}}.
}
\examples{
\dontrun{
setProjectLines(simplace, "1,3,6,9-17,33")
setProjectLines(simplace, c(1,2,3,9:17,33))}
}
