% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{setSimplaceDirectories}
\alias{setSimplaceDirectories}
\title{Set working-, output-, projects- and data-directory}
\usage{
setSimplaceDirectories(
  simplace,
  WorkDir = nullString,
  OutputDir = nullString,
  ProjectsDir = nullString,
  DataDir = nullString
)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}

\item{WorkDir}{working directory where solutions, projects and data resides (_WORKDIR_)}

\item{OutputDir}{directory for output (_OUTPUTDIR_)}

\item{ProjectsDir}{optional directory for project data  (_PROJECTSDIR_)}

\item{DataDir}{optional directory for data (_DATADIR_)}
}
\value{
No return value, called for the side effect of setting framework directories
}
\description{
One can specify all or only some of the directories. 
Only the directories specified will be set.
}
\seealso{
\code{\link{getSimplaceDirectories}}
}
