% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{setSlotCount}
\alias{setSlotCount}
\title{Sets number of used CPUs}
\usage{
setSlotCount(count)
}
\arguments{
\item{count}{number of processors}
}
\description{
Sets the number of processors that are used parallel.
The function can be used only after \code{\link{initSimplace}} has been
called.
}
