% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{stepSimulation}
\alias{stepSimulation}
\title{Run simulation stepwise}
\usage{
stepSimulation(
  simplace,
  count = 1,
  filter = NULL,
  parameterList = NULL,
  simulationNumber = 1
)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}

\item{count}{number of steps to be performed}

\item{filter}{vector of the variable names to be included in the result. If not set, all variables are returned}

\item{parameterList}{list of parameter values indexed by parameter name}

\item{simulationNumber}{number of simulation in the queue that should be run stepwise (default first simulation)}
}
\value{
handle to the data container which has to be processed afterwards
}
\description{
Performs \code{count} steps of the simulation and returns the values from 
the actual variable map. Can be called consecutively.
}
\examples{
\dontrun{
simplace <- initSimplace(SimplaceInstallationDir,SimplaceWorkDir,SimplaceOutputDir)
openProject(simplace, Solution)
createSimulation(simplace)
vm <- stepSimulation(simplace,count=22)
vm_s <- stepSimulation(simplace,filter=c("CURRENT.DATE","LintulBiomass.sWSO"),count=18)
closeProject(simplace)   }

}
