% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{openProject}
\alias{openProject}
\title{Opens a Simplace project}
\usage{
openProject(simplace, solution, project = nullString)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}

\item{solution}{solution file with absolute path}

\item{project}{project file with absolute path, can be omitted to run solution only}
}
\description{
Initializes a project. The absolute path to a solution file is mandatory. 
Project file is optional.
}
\seealso{
\code{\link{closeProject}}
}
