% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{openProject}
\alias{openProject}
\title{Opens a Simplace project}
\usage{
openProject(simplace, solution, project = nullString, parameterList = NULL)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}

\item{solution}{solution file with absolute path or path relative to workdir}

\item{project}{project file with absolute path or path relative to workdir, can be omitted to run solution only}

\item{parameterList}{a list with the parameter name as key and parametervalue as value}
}
\description{
Initializes a project. The absolute path to a solution file is mandatory. 
Project file is optional.
}
\seealso{
\code{\link{closeProject}}
}
